Initial import of the client side launcher application for Google Chrome
This will provide a simple launcher script that will be distributed along
with Google Chrome's flatpak, to be called from the OS-level system helper
app, which knows how to run Chrome from outside the flatpak.

Doing this would allow us to update versions of this launcher script when
new versions of Chrome come if needed, as well as to adapt in the future
to the new mechanism for external apps (using flatpak's upstream preferred
method) without requiring OS updates, as the system helper will remain
as simple as possible.

https://phabricator.endlessm.com/T14244