Update to org.freedesktop.Platform//18.08
This app doesn't actually use its runtime (except when deploying
extra-data) since it is run on the host system. However, we want it to
use a runtime that is included in all Endless OS images and is likely to
be widely installed. Now that LibreOffice (which is pre-installed on all
systems) uses the 18.08 version of the platform, this app can/should use
it too.

https://phabricator.endlessm.com/T25269