Fix support for creating desktop shortcuts
When creating desktop shortcuts chrome uses the value of the
CHROME_WRAPPER env variable to build the .desktop file Exec line.
The problem is that the default wrapper script shipped inside the
google-chrome extra-data package sets this variable to the absolute
path of the wrapper script itself which is tied to a specific
version of this flatpak and will change on new updates and break
existing shortcuts.

Let's rename/replace the original binary with another wrapper that
sets CHROME_WRAPPER to our system launcher before invoking the real
chrome binary.

https://phabricator.endlessm.com/T21357