Check for updates using GitHub Actions
Previously, our Jenkins ran a Jenkinsfile in
https://github.com/endlessm/flatpak-external-data-checker which used a
wrapper script which checked a bunch of repos.

Today, the checker has moved to Flathub and only one Endless-specific
app remains to check: com.google.Chrome.

Run that same checker using GitHub Actions, removing the need to
maintain the Endless-specific Jenkinsfile and wrapper script in f-e-d-c
and making the whole thing a bit clearer. The workflow runs on a similar
schedule (every 6 hours) and can be manually triggered.

https://phabricator.endlessm.com/T32493